+++ src/components/home/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SearchBar } from './SearchBar';
 
-export default function MainPage(props) {
+export default function Home(props) {
 
 
     return (
